+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## [Unreleased]
 
+## [0.10.0] - 2020-05-19
 ### Added
 - Dynamic type webhooks without the need to a specific type (can use as multitype webhook).
 
@@ -81,7 +82,8 @@ Breaking: Webhook constructors now need a tracer.
 - Static mutating webhook.
 - Handler creator for webhooks.
 
-[Unreleased]: https://github.com/slok/kubewebhook/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/slok/kubewebhook/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/slok/kubewebhook/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/slok/kubewebhook/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/slok/kubewebhook/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/slok/kubewebhook/compare/v0.7.0...v0.8.0